requirement.txt
torch
transformer
streamlit
assemblyai